PMM-6324 Refactor table components parameters
