+++ pmm-app/src/pmm-qan/panel/components/Details/Table/components/Table.types.ts
@@ -1,0 +1,7 @@
+import { DatabasesType } from '../../Details.types';
+
+export interface TableProps {
+  tableName: string;
+  databaseType: DatabasesType;
+  example: any;
+}
